4.21*

+++ Osa4/blogilista/controllers/blogs.js
@@ -1,32 +1,51 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})  
-  response.json(blogs)
-})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
+  console.log(blogs)
+  response.json(blogs);
+});
   
-blogsRouter.post('/',  async (request, response) => {
+
+blogsRouter.post('/', userExtractor, async (request, response) => {
     const body = request.body 
-    console.log(body)
+    // get user from request object
+    const user = request.user
+
     const blog = new Blog({
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes
+      likes: body.likes,
+      user: user._id
     })
     if (!blog.title || !blog.url){
-      response.status(400).end()
+      response.status(400).json({ error: 'title or ulr missing'})
     }
     else {
       const savedBlog = await blog.save()
+      user.blogs = user.blogs.concat(savedBlog._id)
+      await user.save()
       response.status(201).json(savedBlog)
     }
     
   })
-blogsRouter.delete('/:id', async (request,response)=> {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request,response)=> {
+
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  console.log(blog)
+  console.log(user)
+  if (blog.user.toString() === user.id.toString()){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  
+  response.status(400).json({ error: 'token invalid'})
 })
 
 blogsRouter.put('/:id', async (request,response) => {
